+++ CHANGELOG.markdown
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+* Unify getting description of option default value using `default_description` [@toy](https://github.com/toy)
 * Don't use `-strip` option for optipng when the bin version is less than 0.7 [#106](https://github.com/toy/image_optim/issues/106) [@toy](https://github.com/toy)
 * Use quality `0..100` by default in lossy mode of pngquant worker [#77](https://github.com/toy/image_optim/issues/77) [@toy](https://github.com/toy)